+++ src/main/kotlin/presentation/uiController/MainMenuHandler.kt
@@ -16,14 +16,6 @@ class MainMenuHandler(
                     println("See You Later 🙂")
                     break
                 }
-                3 -> {
-
-                }
-
-                15 -> {
-
-                }
-
                 null -> println("❌ Invalid input. Please enter a valid number.")
 
                 else -> println("❌ Unknown feature number.")
